% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DtMat.R
\name{Dt}
\alias{Dt}
\title{Calculate Dt Matrix}
\usage{
Dt(t, h, FixT)
}
\arguments{
\item{t}{a new time position}

\item{h}{bandwidth}

\item{FixT}{fixed time positions}
}
\value{
a matrix
}
\description{
Calculate Dt matrix to be used in calculating smoothing matrix.
}
