% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AfterProfile.R
\name{est}
\alias{est}
\title{Estimate beta, sigma2, f(t) and covariance matrix}
\usage{
est(theta.best, h, st, KERN1, KERN2, CovStructure, loctim, LocDistMat,
  TimeDistMat, P, X, Z, Smat, Gmat, Sigma = NUL)
}
\arguments{
\item{theta.best}{profile likelihood estimator of theta (not including sigma2)}

\item{CovStructure}{specify the covariance structure, 1 - a separable covariance matrix, 2 - a nonseparable covariance matrix}

\item{LocDistMat}{the distance matrix of location}

\item{TimeDistMat}{the distance matrix of time}

\item{P}{the dimension of data}

\item{X}{the design matrix of linear part of the model}

\item{Z}{the dependent variable}

\item{Smat}{the smoothing matrix}
}
\value{
\item{beta}{Linear regression paramters estimation}
\item{theta}{Covariance paramters estimation}
\item{fest}{f(t) estimation}
\item{Sigma}{Estimated covariance matrix}
\item{J0theta}
\item{J0betainv}
\item{J0f}
}
\description{
Estimate beta, sigma2, f(t) and covariance matrix after obtaining the profile likelihood estimator of theta (sigma2 excluded).
}
\keyword{likelihood}
\keyword{profile}
