% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bimodal.R
\name{K2}
\alias{K2}
\title{Bimodal Kernel Function}
\usage{
K2(u)
}
\arguments{
\item{u}{a value}
}
\value{
the value of bimodal kernel function at u
}
\description{
Calculate the value of a bimodal kernel function
}
