% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KtMat.R
\name{Kt}
\alias{Kt}
\title{Calculate Kt Matrix in PLM}
\usage{
Kt(t, h, FixT, KN)
}
\arguments{
\item{t}{a new time position}

\item{h}{bandwidth}

\item{FixT}{fixed time positions}

\item{KN}{Gaussian kernel or bimodal kernel}
}
\value{
a matrix
}
\description{
Calculate Kt matrix to be used in calculating the smoothing matrix
}
