% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateProfile.R
\name{log.profile}
\alias{log.profile}
\title{Calculate the Profile Likelihood}
\usage{
\method{log}{profile}(theta0, CovStructure, loctim, LocDistMat, TimeDistMat, P,
  Xprime, Zprime, Sigma = NULL)
}
\arguments{
\item{theta0}{initial covariance paramters (not including sigma2)}

\item{CovStructure}{the covariance structure, 1 - a separable covariance matrix, 2 - a nonseparable covariance matrix}

\item{LocDistMat}{the distance matrix of location}

\item{TimeDistMat}{the distance matrix of time}

\item{P}{the dimension of data}

\item{Xprime}{transformed X, Xprime = (I-S)X}

\item{Zprime}{transformed Z, Zprime = (I-S)Z}
}
\value{
the value of log profile likelihood.
}
\description{
Calculate the Profile Likelihood based on initial covariance parameters (not including sigma2)
}
