% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLMfitFixedBandwidth.R
\name{spl}
\alias{spl}
\title{Fitting Spatio-Temporal Model with A Specified Bandwidth}
\usage{
spl(theta0, h, st, KERN1, KERN2, CovStructure, loctim, LocDistMat, TimeDistMat,
  P, X, Z, Sigma = NULL, neigh = Inf)
}
\arguments{
\item{theta0}{initial covariance paramters (not including sigma2)}

\item{h}{the prespecified bandwidth}

\item{st}{the index of kernel}

\item{KERN1, KERN2}{two candidate kernels}

\item{CovStructure}{the covariance structure, 1 - a separable covariance matrix, 2 - a nonseparable covariance matrix}

\item{loctim}{(s,t) the spatiotemporal index}

\item{LocDistMat}{the distance matrix of location}

\item{TimeDistMat}{the distance matrix of time}

\item{P}{the dimension of data}

\item{X}{X}

\item{Z}{y}

\item{Sigma}{The true covariance matrix}

\item{neigh}{the range of parameter estimation (Inf for stationary case)}
}
\value{
beta.hat, theta.hat, f.hat, Sigma.hat, profile likelihood
}
\description{
Fit a spatio-temporal model with a specified bandwidth
}
