% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmoothMat.R
\name{St}
\alias{St}
\title{Calculate the Smoothing Matrix}
\usage{
St(t, h, FixT, KN)
}
\arguments{
\item{t}{a new time position}

\item{h}{bandwidth}

\item{FixT}{fixed time positions}

\item{KN}{Gaussian kernel or bimodal kernel}
}
\value{
the smoothing matrix
}
\description{
Calculate the Smoothing Matrix
}
